jsdocs on mock objects and move helper function to gameManager.ts
