+++ src/test/utils/gameManager.ts
@@ -138,6 +138,38 @@ export default class GameManager {
   }
 
   /**
+   * Helper function to run to the final boss encounter as it's a bit tricky due to extra dialogue
+   * Also handles Major/Minor bosses from endless modes
+   * @param game - The game manager
+   * @param species
+   * @param mode
+   */
+  async runToFinalBossEncounter(game: GameManager, species: Species[], mode: GameModes) {
+    console.log("===to final boss encounter===");
+    await game.runToTitle();
+
+    game.onNextPrompt("TitlePhase", Mode.TITLE, () => {
+      game.scene.gameMode = getGameMode(mode);
+      const starters = generateStarter(game.scene, species);
+      const selectStarterPhase = new SelectStarterPhase(game.scene);
+      game.scene.pushPhase(new EncounterPhase(game.scene, false));
+      selectStarterPhase.initBattle(starters);
+    });
+
+    game.onNextPrompt("EncounterPhase", Mode.MESSAGE, async () => {
+      // This will skip all entry dialogue (I can't figure out a way to sequentially handle the 8 chained messages via 1 prompt handler)
+      game.setMode(Mode.MESSAGE);
+      const encounterPhase = game.scene.getCurrentPhase() as EncounterPhase;
+
+      // No need to end phase, this will do it for you
+      encounterPhase.doEncounterCommon(false);
+    });
+
+    await game.phaseInterceptor.to(EncounterPhase, true);
+    console.log("===finished run to final boss encounter===");
+  }
+
+  /**
    * Transitions to the start of a battle.
    * @param species - Optional array of species to start the battle with.
    * @returns A promise that resolves when the battle is started.
